fix unsigned short

+++ FirstPass.h
@@ -20,11 +20,11 @@
 
 bool isLabel(char* line);
 void firstPass(char *inFile);
-int getLabelValue(char* label);
+unsigned short getLabelValue(char* label);
 void breakLine(char* line, char** label, char** operation, char** datatype, char** args);
 // bool isInTable(char* label);
 void secondPass(char *inFile);
-int intToOctal(int num);
+unsigned short intToOctal(unsigned short num);
 void entryFileMaker(char* inFile);
 void externFileMaker(char* inFile);
 void objectFileMaker(char* inFile);
